# openapi.yaml
openapi: 3.0.0
info:
  title: Pet Adoption System API
  description: API for user registration, login, adopter preferences, shelter pet
    management, and recommendations in the Pet Adoption System.
  version: 5.0.0
  contact:
    name: PRJ666 - Group 1 - 'Pet Adoption System' Team
    email: support@petadoption.com
servers:
  - url: ${openapi.petAdoptionSystem.base-path}
    description: Configurable base path for the API
...
paths:
...
  /matching/recommendations:
    get:
      tags:
        - matching
      summary: Get a list of recommended pets based on preferences
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: A list of recommended pets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          $ref: '#/components/schemas/Recommendation'
  /matching/recommendation/next:
    get:
      tags:
        - matching
      summary: Get the next recommendation for the adopter
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: The next recommendation item
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Recommendation'
  /matching/recommendation/{id}/accept:
    put:
      tags:
        - matching
      summary: Accept a recommendation and create an adoption
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The ID of the recommendation to accept
      responses:
        '201':
          description: The next recommendation item
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: string
                        example: "adfa4531..."
  /matching/recommendation/{id}/reject:
    put:
      tags:
        - matching
      summary: Reject a recommendation
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The ID of the recommendation to reject
      responses:
        '201':
          description: Recommendation rejected
          $ref: '#/components/responses/GenericApiResponse'
  /adoption/status:
    get:
      tags:
        - adoption
      summary: Retrieve adoption status
      security:
        - BearerAuth: [ ]
      description: Fetches the current status of adoptions for the authenticated adopter.
      responses:
        '200':
          description: Adoption status retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          type: object
                          properties:
                            adoptionId:
                              type: string
                              example: "adf23ee..."
                            petId:
                              type: string
                              example: "adf23ee..."
                            status:
                              $ref: '#/components/schemas/AdoptionStatus'
                            date:
                              type: string
                              format: date
                              example: '2023-05-15'
        '401':
          description: Unauthorized access
          $ref: '#/components/responses/GenericApiResponse'
  /adoption/story:
    get:
      tags:
        - adoption
      summary: Retrieve a list of adoption stories
      description: Returns a list of adoption stories with the specified number of
        objects.
      parameters:
        - in: query
          name: N
          schema:
            type: integer
            minimum: 1
            default: 5
          description: The number of adoption story objects to return
      responses:
        '200':
          description: A list of adoption stories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          type: object
                          properties:
                            author:
                              type: string
                              description: The author of the adoption story
                              example: John Doe
                            title:
                              type: string
                              description: The title of the adoption story
                              example: A Happy Ending for Max
                            text:
                              type: string
                              description: The content of the adoption story
                              example: Max was found wandering the streets and was taken in
                                by a local shelter. After weeks of care, he found a loving
                                home with the Doe family.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but helps to specify the token type
  responses:
    GenericApiResponse:
      description: Generic API response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
...
    RecommendationStatus:
      type: string
      enum:
        - new
        - accepted
        - rejected
      example: new
    AdoptionStatus:
      type: string
      enum:
        - new
        - pending
        - approved
        - rejected
      example: new
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
          example: "adf12eafd770..."
        petId:
          type: string
          example: "adf12eafd770..."
        date:
          type: string
          format: date
          example: '2023-05-15'
        status:
          $ref: '#/components/schemas/RecommendationStatus'